Rework formats.
Pointers to each format are available globally.  Instead of having
separate "lame", "mad", and "id3" formats, there's one "rsvc_mp3" object
which points to the different implementations.  This is a regression for
m4a, which formerly supported encoding to ALAC, but now only supports
AAC (under the name "m4a").